Extract create option functions
